requirements.txt
async-generator==1.10
attrs==22.2.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
et-xmlfile==1.1.0
##Flask==1.1.2
Flask>=2.2.2
Flask-Cors==3.0.10
greenlet==2.0.2
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
MouseInfo==0.1.3
mysql-connector-python==8.0.32
numpy==1.24.0
openpyxl==3.0.10
outcome==1.2.0
packaging==22.0
pandas==1.5.2
Pillow==9.3.0
protobuf==3.20.3
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
PyMySQL==1.0.2
PyQt5-sip==12.11.0
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytweening==1.0.4
pytz==2022.7
requests==2.28.1
rubicon-objc==0.4.3
selenium==4.7.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==2.0.7
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.5.0
urllib3==1.26.13
webdriver-manager==3.8.5
Werkzeug==2.2.2
wsproto==1.2.0
